+++ compiler/src/lexer.ts
@@ -114,7 +114,6 @@ export class Lexer {
             }
             switch (ident) {
                 case 'let': return { type: TokenType.Let, value: ident, line: this.line };
-                case 'fn': return { type: TokenType.Fn, value: ident, line: this.line };
                 case 'return': return { type: TokenType.Return, value: ident, line: this.line };
                 case 'if': return { type: TokenType.If, value: ident, line: this.line };
                 case 'else': return { type: TokenType.Else, value: ident, line: this.line };